Core: Remove the horrible const_cast hack for the memory view
Memory: Remove the const_cast from ConstMemoryView to MemoryView
Decoder: Adapt accordingly to the memory change

I initially made this horrible hack with const_cast just to avoid
changing the signature of all the decoding functions.
That's bad because it throws away all the const safety (even leading to
possible undefined behaviour); this commit fixes that.