Parser: Fix bug with label checking
While refactoring out a function, I accidentally defined the wrong
condition for an if (with a constexpr bool that was always false).
This made so that the labels would never be verified by the parser, so
the compiler would encode missing labels with zeros and still compile
the (wrong) code.