Parser: Add UnexpectedSectionContent parsing error
momiji-as: Add error handling for the new type
momiji-gl: Add error handling for the new type
momiji-qt: Add error handling for the new type

UnexpectedSectionContent is used when declaring variables in a code section or
when trying to use instructions in a data section. None of these
scenarios actually make sense in a well-formed program, so the parser
should complain about that.
This is also useful to enforce sections for the compiler: since this
enforces a well-formed program, I can use this in the compiler to
shuffle around data and move declarations at the bottom of the
executable memory (just to avoid illegal instruction access).