Parser: Recognize directives and check for invalid sections
Parser: Fix weird variable name
Parser: Sprinkle noexcept and constexpr on trivial functions
Parser: Remove useless stub code

The parser can now detect directives and check for misused sections.
It can tell if a data section contains code or if a code section
contains data declarations. In case of error, the parsing result will be an
UnexpectedSectionContent.

There was a variable named "notdone" inside SeqNext(), the name was not
obvious (I can't even remember why I named it that way) so I changed
that and the code around it to better express my intent.

Before this commit, I stubbed some possible parsing functions for
directives, but it was just gonna make code harder to reason about;
instead I now treat parsing directives as instructions (which should not be a
problem for the compiler anyway).
I'm not sure about this, I may keep it because it may help with introspecting
the parsed code.