Instructions: Initial work on handling status register updates by common instructions
As of now, the only instruction which updates the status register is "cmp", no
other instruction does that.
This commit provides initial work on supporting status register updates
with other instructions (in this case, "move", "add" and "sub").
I still don't know how to detect a carry, so that is unimplemented.