momiji-qt: Add an help window to see help files
momiji-qt: Add stub help files
momiji-qt: Rearrange resource file
momiji-qt: Add initial icons for the main window
gitignore: Ignore Krita's temporary files

This commits implements an help window and related help files.
This is useful if the user wants to know the status of implementation of
the various instructions (or just the terminology used in the project).

Other than that, there are some icons for the main window buttons (along
with the Krita project files).

The documentation is still work in progress, unfortunately.