Core: Fix incorrect byte memory reads when reading a source operand
momiji-gl: Fix incorrect memory visualization
momiji-dump: Fix incorrect memory visualization

When reading a source operand for the instruction with the byte data type, the
interpreter would read the wrong byte (with an offset of PC + 3, while
it should have been PC + 2), reading the padding byte (to align memory
to 2 bytes) which is zero.
That meant that a `move.b` would always set to zero the destination
operand.

Also, this commit addresses incorrect visualization of memory in the
OpenGL and CLI tools, where the order of bytes would be displayed in
reverse.

All of this was caused by an older implementation of `BasicMemory` where
bytes would be (erroneously) added in reverse, so applications needed to
be adjusted accordingly. That problem was fixed long ago, but I forgot to
actually change the applications to reflect that.