Core: Make Register a regular type
Before this commit, the Register was kinda useless: it just provided a `value`
member and that's it.
Now a Register is type-safe: it encapsulates the IntType that it holds and it
behaves like a value.

This will be useful in the future when I'll work on a way to detect
carries and generally avoid UB in low-level operations.

Right now it provides a function to reinterpret the stored value as a pointer
(scary stuff) and another one to get a copy of the stored value.