Emulator: Replace rollback() with rollbackSys()
Core: Remove an old, useless, header

rollback() is now replaced by rollbackSys(), which means that rollbackSys() got
renamed to rollback()!

When the two functions were separate, rollback() was basically dead code: it
was old stuff left behind from the times when the emulator would work on whole
instructions and not raw binary data: there was no decoder back then and the
program counter would index an array of instructions. In those days, "rollback"
meant decrementing the program counter by 1 without touching the state in which
the system was.
Now the whole emulator works in a totally different way that makes this
unfeasible, so it was just dead code and users were encouraged to call
rollbackSys() anyway. With this commit, there's no more confusion about what
"rollback" means.